Adds cost solvers for transactions
